add appllication

+++ openwsn-fw/bsp/boards/OpenMote-CC2538/radio.c
@@ -5,6 +5,9 @@
  * Description: CC2538-specific definition of the "radio" bsp module.
  */
 
+#include <stdio.h>
+#include <string.h>
+
 #include <headers/hw_ana_regs.h>
 #include <headers/hw_ints.h>
 #include <headers/hw_rfcore_sfr.h>
@@ -12,17 +15,15 @@
 #include <headers/hw_rfcore_xreg.h>
 #include <headers/hw_types.h>
 
+#include <source/interrupt.h>
+#include <source/sys_ctrl.h>
+
 #include "board.h"
-#include "radio.h"
-#include "leds.h"
-#include "stdio.h"
-#include "string.h"
-#include "radiotimer.h"
-#include "debugpins.h"
-#include "interrupt.h"
-#include "sys_ctrl.h"
-#include "sys_ctrl.h"
 #include "cc2538rf.h"
+#include "debugpins.h"
+#include "leds.h"
+#include "radio.h"
+#include "radiotimer.h"
 
 //=========================== defines =========================================
 
@@ -251,7 +252,7 @@ void radio_rfOff() {
 
 //===== TX
 
-void radio_loadPacket(uint8_t* packet, uint8_t len) {
+void radio_loadPacket(uint8_t* packet, uint16_t len) {
    uint8_t i=0;
    
    // change state
